+++ Controls/_tile/TileView/TileViewModel.ts
@@ -350,63 +350,6 @@ var TileViewModel = ListViewModel.extend({
         return itemWidth ? Math.max(resultWidth, itemWidth) : resultWidth;
     },
 
-    shouldOpenExtendedMenu(isActionMenu: boolean, isContextMenu: boolean, itemData: Record<string, any>): boolean {
-        const isScalingTile = this._options.tileScalingMode !== 'none' &&
-                              this._options.tileScalingMode !== 'overlap' &&
-                              !itemData.dispItem.isNode();
-        return this._options.actionMenuViewMode === 'preview' && !isActionMenu && !(isScalingTile && isContextMenu);
-    },
-
-    getActionsMenuConfig(
-        itemData,
-        clickEvent: SyntheticEvent,
-        opener,
-        templateOptions,
-        isActionMenu,
-        isContextMenu
-    ): Record<string, any> {
-        if (this.shouldOpenExtendedMenu(isActionMenu, isContextMenu, itemData)) {
-            const MENU_MAX_WIDTH = 200;
-            const menuOptions = templateOptions;
-            /* TODO Вынести этот код из модели в контрол плитки
-               https://online.sbis.ru/opendoc.html?guid=7f6ac2cf-15e6-4b75-afc6-928a86ade83e */
-            const itemContainer = clickEvent.target.closest('.controls-TileView__item');
-            const imageWrapper = itemContainer.querySelector('.controls-TileView__imageWrapper');
-            if (!imageWrapper) {
-                return null;
-            }
-            let previewWidth = imageWrapper.clientWidth;
-            let previewHeight = imageWrapper.clientHeight;
-            menuOptions.image = itemData.imageData.url;
-            menuOptions.title = itemData.item.get(itemData.displayProperty);
-            menuOptions.additionalText = itemData.item.get(templateOptions.headerAdditionalTextProperty);
-            menuOptions.imageClasses = itemData.imageData?.class;
-            if (this._options.tileScalingMode === TILE_SCALING_MODE.NONE) {
-                previewHeight = previewHeight * ZOOM_COEFFICIENT;
-                previewWidth = previewWidth * ZOOM_COEFFICIENT;
-            }
-            menuOptions.previewHeight = previewHeight;
-            menuOptions.previewWidth = previewWidth;
-
-            return {
-                templateOptions,
-                closeOnOutsideClick: true,
-                maxWidth: menuOptions.previewWidth + MENU_MAX_WIDTH,
-                target: imageWrapper,
-                className: 'controls-TileView__itemActions_menu_popup',
-                targetPoint: {
-                    vertical: 'top',
-                    horizontal: 'left'
-                },
-                opener,
-                template: 'Controls/tile:ActionsMenu',
-                actionOnScroll: 'close'
-            };
-        } else {
-            return null;
-        }
-    },
-
     getRoundBorderClasses(): string {
         const theme = `_theme-${this._options.theme}`;
         const topLeftBorder = this._options.roundBorder?.tl || 'default';
