+++ Controls/_tile/View.ts
@@ -2,6 +2,7 @@ import {View as List} from 'Controls/list';
 import {TreeControl} from 'Controls/tree';
 import TreeTileViewModel = require('Controls/_tile/TreeTileView/TreeTileViewModel');
 import TreeTileView = require('Controls/_tile/TreeTileView/TreeTileView');
+import {TILE_SCALING_MODE, ZOOM_COEFFICIENT} from 'Controls/_tile/TileView/resources/Constants';
 
 'use strict';
 
@@ -36,7 +37,7 @@ import TreeTileView = require('Controls/_tile/TreeTileView/TreeTileView');
  *
  * @mixes Controls/_list/interface/IVirtualScrollConfig
  *
- * 
+ *
  * @author Авраменко А.С.
  * @public
  */
@@ -67,7 +68,7 @@ import TreeTileView = require('Controls/_tile/TreeTileView/TreeTileView');
  *
  * @mixes Controls/_list/interface/IVirtualScrollConfig
  *
- * 
+ *
  * @author Авраменко А.С.
  * @public
  */
@@ -78,6 +79,60 @@ export default class View extends List {
    protected _supportNewModel: boolean = false;
    protected _beforeMount(): void {
       this._viewModelConstructor = this._getModelConstructor();
+   }
+
+   private _getItemDataByItem(item): Record<string, any> {
+      //TODO: newModel this._collection.getItemBySourceIndex();
+      return this._children.listControl._children.baseControl._listViewModel.getItemDataByItem(item);
+   }
+   private _shouldOpenExtendedMenu(isActionMenu, isContextMenu, item): boolean {
+      const isScalingTile = this._options.tileScalingMode !== 'none' &&
+          this._options.tileScalingMode !== 'overlap' ||
+          item.isNode();
+      return this._options.actionMenuViewMode === 'preview' && !isActionMenu && !(isScalingTile && isContextMenu);
+   }
+
+   protected _getActionsMenuConfig(e, item, clickEvent, action, isContextMenu, defaultConfig): Record<string, any> {
+      const isActionMenu = !!action && !action.isMenu;
+      if (this._shouldOpenExtendedMenu(isActionMenu, isContextMenu, item)) {
+         const MENU_MAX_WIDTH = 200;
+         const menuOptions = defaultConfig.templateOptions;
+         const itemData = this._getItemDataByItem(item);
+         const itemContainer = clickEvent.target.closest('.controls-TileView__item');
+         const imageWrapper = itemContainer.querySelector('.controls-TileView__imageWrapper');
+         if (!imageWrapper) {
+            return null;
+         }
+         let previewWidth = imageWrapper.clientWidth;
+         let previewHeight = imageWrapper.clientHeight;
+         menuOptions.image = itemData.imageData.url;
+         menuOptions.title = itemData.item.get(itemData.displayProperty);
+         menuOptions.additionalText = itemData.item.get(menuOptions.headerAdditionalTextProperty);
+         menuOptions.imageClasses = itemData.imageData?.class;
+         if (this._options.tileScalingMode === TILE_SCALING_MODE.NONE) {
+            previewHeight = previewHeight * ZOOM_COEFFICIENT;
+            previewWidth = previewWidth * ZOOM_COEFFICIENT;
+         }
+         menuOptions.previewHeight = previewHeight;
+         menuOptions.previewWidth = previewWidth;
+
+         return {
+            templateOptions: menuOptions,
+            closeOnOutsideClick: true,
+            maxWidth: menuOptions.previewWidth + MENU_MAX_WIDTH,
+            target: imageWrapper,
+            className: 'controls-TileView__itemActions_menu_popup',
+            targetPoint: {
+               vertical: 'top',
+               horizontal: 'left'
+            },
+            opener,
+            template: 'Controls/tile:ActionsMenu',
+            actionOnScroll: 'close'
+         };
+      } else {
+         return null;
+      }
    }
    protected _getModelConstructor() {
       return TreeTileViewModel;
